predictions/services/scraper.py
"""
Scraper Service for SmartAcca
Scrapes tipster predictions from multiple sources
"""
import requests
from bs4 import BeautifulSoup
from datetime import datetime, timedelta
import logging

logger = logging.getLogger(__name__)


class TipsterScraper:
    """
    Scrapes tipster predictions from various football prediction websites.
    Currently supports placeholder URLs for Tipstrr, Squawka, and PredictZ.
    """

    def __init__(self):
        self.sources = {
            'tipstrr': 'https://www.tipstrr.com/todays-tips',
            'squawka': 'https://www.squawka.com/predictions',
            'predictz': 'https://www.predictz.com/predictions/today/'
        }
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }

    def scrape_all(self):
        """
        Scrape all tipster sources and return combined results.

        Returns:
            list: List of dictionaries containing match data and tipster picks
        """
        all_predictions = []

        for source_name, url in self.sources.items():
            try:
                predictions = self._scrape_source(source_name, url)
                all_predictions.extend(predictions)
                logger.info(f"Successfully scraped {len(predictions)} matches from {source_name}")
            except Exception as e:
                logger.error(f"Error scraping {source_name}: {str(e)}")
                continue

        # Group predictions by match
        grouped_predictions = self._group_by_match(all_predictions)

        return grouped_predictions

    def _scrape_source(self, source_name, url):
        """
        Scrape a single tipster source.

        Args:
            source_name: Name of the source
            url: URL to scrape

        Returns:
            list: List of predictions from this source
        """
        try:
            response = requests.get(url, headers=self.headers, timeout=10)
            response.raise_for_status()

            soup = BeautifulSoup(response.content, 'html.parser')

            # This is a placeholder implementation
            # In production, you would parse the actual HTML structure
            predictions = self._parse_html(soup, source_name)

            return predictions

        except requests.RequestException as e:
            logger.error(f"Request error for {source_name}: {str(e)}")
            return []

    def _parse_html(self, soup, source_name):
        """
        Parse HTML to extract match predictions.
        This is a placeholder that returns mock data for demonstration.

        Args:
            soup: BeautifulSoup object
            source_name: Name of the source

        Returns:
            list: Parsed predictions
        """
        # PLACEHOLDER: Return mock data for now
        # In production, implement actual HTML parsing logic

        mock_predictions = [
            {
                'home_team': 'Manchester United',
                'away_team': 'Liverpool',
                'pick': 'Over 2.5 Goals',
                'source': source_name,
                'match_date': datetime.now() + timedelta(days=1)
            },
            {
                'home_team': 'Arsenal',
                'away_team': 'Chelsea',
                'pick': 'Home Win',
                'source': source_name,
                'match_date': datetime.now() + timedelta(days=1)
            },
            {
                'home_team': 'Barcelona',
                'away_team': 'Real Madrid',
                'pick': 'Both Teams to Score',
                'source': source_name,
                'match_date': datetime.now() + timedelta(days=2)
            }
        ]

        logger.warning(f"Using mock data for {source_name}. Implement actual parsing logic for production.")

        return mock_predictions

    def _group_by_match(self, predictions):
        """
        Group predictions by match and combine tipster picks.

        Args:
            predictions: List of individual predictions

        Returns:
            list: Grouped predictions with all tipster picks
        """
        match_dict = {}

        for pred in predictions:
            match_key = f"{pred['home_team']}_vs_{pred['away_team']}"

            if match_key not in match_dict:
                match_dict[match_key] = {
                    'home_team': pred['home_team'],
                    'away_team': pred['away_team'],
                    'match_date': pred['match_date'],
                    'tipster_picks': []
                }

            match_dict[match_key]['tipster_picks'].append({
                'source': pred['source'],
                'pick': pred['pick']
            })

        # Convert to list and ensure we have up to 3 tipster picks per match
        grouped = []
        for match_data in match_dict.values():
            picks = match_data['tipster_picks']

            result = {
                'home_team': match_data['home_team'],
                'away_team': match_data['away_team'],
                'match_date': match_data['match_date'],
                'tipster_1_pick': picks[0]['pick'] if len(picks) > 0 else None,
                'tipster_2_pick': picks[1]['pick'] if len(picks) > 1 else None,
                'tipster_3_pick': picks[2]['pick'] if len(picks) > 2 else None,
            }

            grouped.append(result)

        return grouped


def scrape_todays_tips():
    """
    Convenience function to scrape today's tips.

    Returns:
        list: Today's match predictions with tipster picks
    """
    scraper = TipsterScraper()
    return scraper.scrape_all()
