requirements.txt
django>=6.0
requests>=2.32.0
google-genai>=1.56.0
beautifulsoup4>=4.14.0
playwright>=1.57.0
django-environ>=0.12.0
celery>=5.6.0
redis>=7.1.0
python-dateutil>=2.9.0
schedule==1.2.2
